[Ref] Various small fixes for 64-bit compile
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@2594 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ mptrack/Autotune.cpp
@@ -295,7 +295,7 @@ void CAutotuneDlg::OnOK()
 {
 	CDialog::OnOK();
 
-	targetNote = m_CbnNoteBox.GetItemData(m_CbnNoteBox.GetCurSel());
+	targetNote = (int)m_CbnNoteBox.GetItemData(m_CbnNoteBox.GetCurSel());
 	pitchReference = GetDlgItemInt(IDC_EDIT1);
 }
